[product-specification.tex, produc-specification.pdf] extension labels reintroduced
Enumeration of odd numbered list elements hidden for flexibility, made all
numbers arabic numerals, removed parentheses in Style2*, set Mark2 as period.

Motivation: The points are labeled in a specific way to indicate the step
they're an extension of, so that it's easy to refer to them in other steps or in
different parts of the use case. With the introduction of easylist for making
lists in the document, these subtleties were removed, but have now been
reintroduced.