[#49, #50] update Board API, ChessPiecePawn API
Implements #50 and closes #49. Has restructured allPossibleMoves in
ChessPiecePawn in order to take advantage of new methods in Position.

Has added a private method in ChessPiecePawn called `forward`, which returns the
position n steps forward from the pawn's current position, in order to do away
with error-prone and hard to read arithmetic using "pieceColor".

Removed the setPieceNotMoved method in ChessPiecePawn, as it's easier to keep
the last move in Board than it is to update the pawns every time any piece
moves.

Removed the `resetPawnMoves` method in Board, as this is no longer required and
calls on an API that no longer exists.