[ChessPieceKing] fix broken method, small rewrite
Fixed `allEnemyMoves` returning the position of enemies instead of the positions
they can move to.

Rewrote half of `allPossibleMoves` to make use of the new `neighbors` method in
Position and functional programming style. I don't dare to touch the other half
right now, but I think I have to before long.

NOTE: There is a small problem with allEnemyMoves, which is that it calls
allPossibleMoves for all enemy pieces. This, in turn, calls allPossibleMoves for
the enemy king, which means we recurse back into allPossibleMoves for our own
king, and thus we are stuck in infinite recursion. This has been mitigated by
using only the unoccupied neighbors of the enemy king as possible enemy king
moves - however, this is likely to introduce edge cases where legal moves are
disallowed due to the enemy king being seen as having less moves than he does.

For the time being, I will assume this method to return all possible enemy moves
to a first approximation.