[ChessPieceKing] fix castling possible check
Rewrote and fixed check for whether castling is possible. The new approach makes
use of a more functional style, which makes the code more declarative and
readable. Added checks for every square between the castle and king that the
square is not threatened and unoccupied.

The remaining test case in error fails because the code for moving the rook (in
Board) is not working.