[Board.java] fix rook reposition in castling
When castling, the rook is placed one square closer to the center of the board
than the king. Or so I was led to believe. Fight me.

Uses putAtPosition instead of movePiece in order to force the piece to be moved
instead of asking it politely, as movePiece does. Normally the rook would not be
able to move through the king to the given position.

Holy fudge, the AI is broken.