[CastlingMove] fix logic errors
Change old rook position to be rook's position instead of king's.
Change new position of left rook to be east of new pos instead of old pos.
Change new position of king castling with right rook to be east instead of west.
Change new position of right rook to be west of new pos.