[Board] change removePiece code
In a list `a`, if you call `a.remove(elem)`, this shifts the positions in the
lists of the other elements. In order to not invalidate a lot of positions, and
also to try and not use too many resources, the code was changed to do as
follows:

1. If the position of the piece in the list is not equal to the last position,
swap it with the last position and update the position index of that piece
2. Remove the last element and set its position index to -1

This should be fairly efficient (maybe even more than the old code, since we
don't need to delete elements) and should not invalidate any positions.