[Close #75] add dismiss game over overlay button
Added a button to dismiss game overlay, and also added a button to show the
results, and made it so that this button is visible in place of the give up
button if the game is over.

Also changed the turn text as appropriate whenever the game ends. Helper method
setTurnText added.