[Close #78] PlayerAttributes added to GameScreen
Added a helper method in SetupScreen to make the creation of the GameScreen more
readable, but a lot of stuff there should be refactored. Don't know how color
will be selected, so I just created the colors black and white on need for the
time being - that should be changed later. I have added a color attribute to the
GameScreen fields for each of the players, but I have not used it for drawing
the pieces - the color is currently disconnected. I'm not sure exactly how it's
best to connect it, but we can probably use the color directly when setting the
color, if I understand it correctly. However, as for special effects such as
highlighting, I'm not sure.