[Screen classes] refactor
Refactor identical methods to AbstractScreen. This includes dispose, which
required me to also refactor the stage field. That caused a NullPointerException
in AccountOverlay.setStage despite there being no reason why that would happen.
Adding an early return in case of stage being null, the issue was solved
somehow, and the game still seems to work.