[SetupScreen] refactor
Mainly refactored the addDifficultyList method, ended up adding a new method in
order to be able to make use of the new createButton method, and found another
place where it could replace a few lines of code to convince myself it was worth
it. Also, removed switch statement on index to get button text, put them in an
array instead.