[Position] fix bugs in stepsToPosition
Also created Position.Manipulator interface for functional style position
manipulation, and changed ChessPieceKing's Manipulator-using function to use
Position.Manipulator instead of its own private version.

Position.Manipulator can be used when, given a position, you need to determine a
"next" position. A very useful way to use it is as a way to specify a direction
- i.e, you can pass a Manipulator to a function and use it as a way to "keep
stepping in a direction" (for instance, north) until you reach some kind of
criteria. This is the way the Manipulator is used in both ChessPieceKing and in
Position.stepsToPosition.