[Bugfix] invisible AI promotion
Fixed by checking in GameScreen when the AI performs a move, whether the
performed move was a promotion move, in which case it retrieves the piece
currently at the given position and adds it as an actor to the BoardGroup.

Has added a new method `public voidaddPiece(IChessPiece piece)` to make
this possible, and has updated BoardGroup to use it internally.