[#101] Tentatively complete refactor
Important changes that have been made:
- General Game functionality transferred to AbstractGame
- Game renamed to ChessGame
- Previously named ChessGame moved to gui and renamed to LibgdxChessGame
- UI previously having Game (ChessGame) fields have been changed to take
  AbstractGame instead

Also made a lot of other changes required for the code to compile. The tests are
passing (except for hard AI speed test, which was for some reason slow for me
this time), but the game has not been tested to assure that all the GUI stuff
still works as expected. There are also some other changes that probably need to
be done, but for now, the project compiles again.