[90e4d59] Fix move history crash, close #81
The move history was broken when I removed the old, outdated versions of methods
which AbstractGame was supposed to reimplement, due to the fact that the old
version computes the added move when it was made, whereas the AbstractGame
version that was based on the old one computes this for every move every time
the history changes. Thus, the move history relies on old data, and it turned
out that this quickly resulted in a NullPointerException.

The fix overrides Move.toString which returns a string representing the move
which is valid regardless of the current board state. It was decided to use long
algebraic notation because it is easier to generate based on the information we
currently have in the Move class.

The toString methods of PromotionMove and CastlingMove were also overridden, and
EnPassantMove was changed in order to let its superclass know of the piece it
takes.

Additionally, this commit changes ChessPiecePawn.toChessNotation to return an
empty string (since this is usually what's used to represent pawns in AN), and
Position got the new utility method toChessNotation which returns a string on
the form of [a-h][1-8]. The move numbering was also changed to be based on the
total number of moves done, as is usually the case in chess notation.