[#115] Tentative revamp of SetupScreen
Added new RadioButton extends Actor and RadioGroup extends Group, and modified
SetupScreen to create and use RadioGroups for both players, and modified
startGame so that game is still started with correct parameters.

The UI should probably be tweaked a little. For one, maybe we want to fix the
heavy distortion by WobblyDrawable, and we may want to put the bottom buttons
further down, or at least add some more setup options in between, because right
now it looks a bit squashed.

SetupScreen should probably also get a refactoring, since there are a lot of
things which aren't used anymore.