[#115] Refactor more and add CheckButton
Refactored selectable button behavior (button with some icon indicating
selection, text next to the button describing what it's for, etc.) into
AbstractSelectableButton, and let the subclasses specify the textures to
indicate selected/unselected state (with the checked texture on top of the box
texture), as well as the selection behavior to perform if the button is
selected. RadioButton sets selected to true no matter what, while CheckButton
toggles.

Now, the button "Play random game" can be eliminated from MainMenuScreen.