[SjadamGame] Update movement algorithms
Updated allPossibleMoves and performMove methods in order to be able to
determine which moves are allowed or not given the state of the board. Also
updated tests to reflect an increased understanding of the ruleset. All tests
passing.

TODO: Override copy and undo methods as appropriate in order to ensure that no
board corruption can occur if the AI utilizes the SjadamGame API.