[#98/#107] Animation update and Game API change
Animation was updated and is now able to animate multiple position changes
simultaneously. It might be better to animate them sequentially, but for now
this is adequate.

The performMove API is increasingly unjustifiable the way it currently works, so
it was replaced with getMoves which returns the same thing but doesn't actually
perform the move, and the performMove(Position, Position) method was set to
deprecated. The UI now picks the move to perform and sees it done with
performMove(Move).

The API update breaks a test case, which needs to be updated.